Fix assignment of 'ret'
The comparison operator (!=) has a higher precedence than the
assignment operator (=), so in all these lines, we were
setting 'ret' to either 0 or 1, the result of the comparison
operator, instead of setting it to the result of the functions.

We use parentheses to instead assign 'ret' to the result of the
functions.

Test: TreeHugger
Bug: 202771600
Change-Id: Ibe91ae4dd5a39e8ef8ac3d4f1639a4581ebc5a43