libhwc2.1: check the overlapped and assigned layers fit constraints
After checking the current layer’s constraints, HWC needs to check the
overlapped layer constraints. When these overlapped layers were
confirming their constraints, the current layer was not assigned to a
MPP, so it was not considered in overlapped layers’s
isHWResourceAvailable(). This causes HWC to assign more layers than
layer constraints.

Bug: 264601428
Test: use test-surface to simulate the b/259317030
Change-Id: I7394015d43c5bbd0f37553a6902715618e8aa5b6